introduce package app and server struct
