+++ app/lepus.go
@@ -1,0 +1,40 @@
+// copyright Jusong Chen
+
+package app
+
+import (
+	"log"
+	"net/http"
+
+	"github.com/go-chi/chi"
+)
+
+const maxUploadSize = 20 * 1024 * 1024 // 20 mb
+
+type server struct {
+	router        *chi.Mux
+	version       string
+	receiveDir    string
+	educatorNames []string
+}
+
+var srv server
+
+var educatorNames = []string{"蔡春耀", "蔡温榄", "蔡小华", "曹世才", "曾敏毅", "沈淑耀", "陈本培", "陈福音", "陈济兴", "陈金治", "陈群青", "陈细英", "陈燕华", "陈由溪", "陈玉珍", "陈长青", "陈志章", "陈宗辉", "傅德卿", "官尚武", "何天祥", "洪大锋", "胡永模", "胡永岳", "江英岩", "蒋永潮", "李粹玉", "练锡康", "林秉松", "林端川", "林芳", "林嘉坚", "林茂英", "林培坚", "林岫英", "林月心", "林占樟", "林昭英", "刘开天", "刘世煌", "刘永宾", "陆佩珰", "罗朝东", "毛玉珍", "毛祖辉", "毛祖瑜", "欧阳兴", "潘宝升", "潘家健", "潘世英", "潘孝平", "钱振恒", "石柏仁", "童美霞", "王福庆", "王光琳", "王丽华", "王美珠", "王其本", "王强", "王如", "王玉芳", "危金炎", "魏友义", "吴大樑", "吴齐练", "吴生基", "肖方尤", "肖光磊", "肖忠波", "许美钗", "严秀凤", "杨诚", "杨虹", "杨孝华", "杨义森", "叶菁", "余春华", "余冬生", "余世棣", "余天雨", "余伟然", "詹玉赐", "张殿", "张桂贞", "张家新", "张孔烺", "张荣治", "张世年", "张芝萱", "赵文婷", "郑春高", "郑国钦", "郑玉珠", "郑宗振", "周治河", "庄可明", "庄瑞发"}
+
+//Start starts Lepus http server
+func Start(port string, dirForStatic string, serverVersion string, receiveDir string) {
+
+	r := chi.NewRouter()
+	srv = server{
+		router:        r,
+		version:       serverVersion,
+		receiveDir:    receiveDir,
+		educatorNames: educatorNames,
+	}
+
+	srv.routes(dirForStatic)
+
+	log.Printf("Server started on port %s", port)
+	log.Fatal(http.ListenAndServe(port, r))
+}
