+++ public/index.html
@@ -17,7 +17,7 @@
     </div>
     <div class="form-group">
         <div class="container-fluid">
-        <form action="/register" method="get" >
+        <form action="/signup" method="get" >
             <input type="submit" value="上传相片或视频">
         </form>
         </div>
