+++ views/signup.html
@@ -1,5 +1,5 @@
 {{define "yield"}}
-<form action="/register" method="post" enctype="application/x-www-form-urlencoded">
+<form action="/signup" method="post" enctype="application/x-www-form-urlencoded">
 
     <div class="form-group">
         <label for="participantName">姓名</label>
